Perform cleanup when WM_QUIT message is read
Perfoming cleanup(destroying window) in tray_exit casues process to be
hang on GetMessage in tray_loop for windows 11. So instead of destroying
window, we post quit message in tray_exit and call cleanup in tray_loop.

Also removing unofficial way to support case sensitive support for
hotkeys. From then on hotkeys are not case sensitive.